feat: use EZE test runner as default environment
Replace the API for test runners with environments. An environment bundles the execution of actions and the access of the process instance state.

Compared to the test runner API, environments are based on Zeebe and doesn't work for a generic BPMN engine. The new API limits the scope of the project (i.e. only Zeebe) but it simplifies the code and the usage by removing abstraction layers.

Use EZE as the default environment for the spec runner because it is very fast and isolated. The environment is based on the previous EZE test runner.