feat: inject spec runner directly in JUnit
Change the usage of the JUnit extension. Instead of injecting a factor for the spec runner, the extension injects the spec runner directly.

The spec runner (i.e. verification timeout/retry-interval) can be configured using the annotation parameters.

This change is a result of removing the test runner API in the core module.